Refactor SimpleHtmlDom in order to make use of namespaces
