+++ .travis.yml
@@ -15,6 +15,9 @@ install:
 before_install:
   - pecl channel-update pecl.php.net
   - pecl install -f mongodb
+  # Ensure there is enough memory to run integration tests (server + client)
+  - echo "memory_limit = 256M" >> ~/.phpenv/versions/$(phpenv version-name)/etc/php.ini
+  - echo "error_reporting = E_ALL & ~E_NOTICE & ~E_STRICT & ~E_DEPRECATED" >> ~/.phpenv/versions/$(phpenv version-name)/etc/php.ini
 script:
   - ./vendor/bin/phpunit
 notifications:
