+++ tests/bootstrap.php
@@ -1,7 +1,12 @@
 <?php
 
-error_reporting(-1);
+//error_reporting(-1);
 date_default_timezone_set('UTC');
+
+ini_set('error_reporting', E_ALL); // or error_reporting(E_ALL);
+ini_set('display_errors', '1');
+ini_set('display_startup_errors', '1');
+
 /*
  * Path trickery ensures test suite will always run, standalone or within
  * another composer package. Designed to find composer autoloader and require
